Devide Position as new type
- Position 자체의 역할도 크다고 생각되어 서브타입에서 별개의 타입으로 나눔
- 테스트를 통해 이 작업이 큰 영향을 주지 않는다는 것을 확인할 수 있었음
